requirements_gpu.txt
wheel
numpy
scipy
scikit-image
tensorflow-gpu
sacred
matplotlib
pymongo
pyqt5